fix(#13): contains the configuration of the server to be accessable
